cleanup: two minor findings from cppcheck
- js_test_engine.c: preprocess_script() wrote the trailing NUL in
  the !match branch right before `break`, then again unconditionally
  after the loop.  Drop the first store; the post-loop one always
  runs (cppcheck redundantAssignment).
- nrf54l15_soc.c: the RADIO SUBSCRIBE_* trace passed `value & 0x1F`
  (uint32_t) into a `%d` slot; add an explicit `(int)` cast
  (cppcheck invalidPrintfArgType_sint).

+++ src/arm/nrf54l15_soc.c
@@ -1449,7 +1449,7 @@ static void nrf54l_radio_write(void *user_data, uint32_t addr, uint32_t value) {
                 "BCSTART","BCSTOP","EDSTART","EDSTOP","CCASTART","CCASTOP"
             };
             fprintf(stderr, "[radio SUBSCRIBE_%s = ch%d cyc=%lld]\n",
-                    idx < 12 ? names[idx] : "?", value & 0x1F,
+                    idx < 12 ? names[idx] : "?", (int)(value & 0x1F),
                     (long long)r->plat->cpu.cycles);
         }
         r->subscribe[idx] = value;
